+++ libs/core/src/lib/date-picker/date-picker.module.ts
@@ -12,6 +12,7 @@ import { FormMessageModule } from '@fundamental-ngx/core/form';
 import { I18nModule } from '@fundamental-ngx/i18n';
 import { DeprecatedDatePickerCompactDirective } from './deprecated-date-picker-compact.directive';
 import { ContentDensityModule } from '@fundamental-ngx/core/content-density';
+import { BarModule } from '@fundamental-ngx/core/bar';
 
 @NgModule({
     declarations: [DatePickerComponent, DeprecatedDatePickerCompactDirective],
@@ -25,6 +26,7 @@ import { ContentDensityModule } from '@fundamental-ngx/core/content-density';
         InputGroupModule,
         ButtonModule,
         FormMessageModule,
+        BarModule,
         ContentDensityModule
     ],
     exports: [DatePickerComponent, DeprecatedDatePickerCompactDirective, ContentDensityModule]